requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
discord.py>=2.3.2
aiohttp>=3.9.0
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.0.0
